+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.3.17 [unreleased]
 - feat: Added Ipfs::pubsub_events to receive subscribe and unsubscribe events to a subscribed topic [PR 70]
+- feat: Implement RepoProvider [PR 69]
 
+[PR 69]: https://github.com/dariusc93/rust-ipfs/pull/69
 [PR 70]: https://github.com/dariusc93/rust-ipfs/pull/70
 
 # 0.3.16